CUSTOM-AGENTS/requirement.txt
langchain>=0.1.17
langgraph>=0.0.24
langchain-google-genai>=0.0.6
python-dotenv>=1.0.1
tiktoken>=0.6.0         # Optional (for token counting/debugging)
streamlit>=1.34.0       # Optional (if you plan to build a UI)
